Move Spy to its own file
